@@ -1,4 +1,4 @@
-name: amaff
+name: hprms2
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
@@ -27,25 +27,26 @@ dependencies:
     sdk: flutter
   auto_size_text: 3.0.0
   cached_network_image: 3.2.1
-  cloud_firestore: 4.2.0
-  cloud_firestore_platform_interface: 5.9.1
-  cloud_firestore_web: 3.1.1
+  cloud_firestore: 4.8.0
+  cloud_firestore_platform_interface: 5.15.0
+  cloud_firestore_web: 3.6.0
   dropdown_button2: 2.0.0
   easy_debounce: 2.0.1
   file_picker: 5.2.6
-  firebase_auth: 4.2.0
-  firebase_auth_platform_interface: 6.11.4
-  firebase_auth_web: 5.2.0
-  firebase_core: 2.4.0
-  firebase_core_platform_interface: 4.5.2
-  firebase_core_web: 2.0.2
-  firebase_storage: 11.0.7
-  firebase_storage_platform_interface: 4.1.25
-  firebase_storage_web: 3.3.17
-  fl_chart: ^0.62.0
+  firebase_auth: 4.6.2
+  firebase_auth_platform_interface: 6.15.2
+  firebase_auth_web: 5.5.2
+  firebase_core: 2.13.1
+  firebase_core_platform_interface: 4.8.0
+  firebase_core_web: 2.5.0
+  firebase_storage: 11.2.2
+  firebase_storage_platform_interface: 4.4.2
+  firebase_storage_web: 3.6.2
   flutter_animate: 4.1.1+1
   flutter_cache_manager: 3.3.0
+  flutter_email_sender: ^6.0.1
   flutter_plugin_android_lifecycle: 2.0.9
+  flutter_spinkit: 5.1.0
   font_awesome_flutter: 10.1.0
   from_css_color: 2.0.0
   go_router: 7.1.1
@@ -61,7 +62,7 @@ dependencies:
   image_picker_ios: 0.8.6+1
   image_picker_platform_interface: 2.6.2
   infinite_scroll_pagination: 3.2.0
-  intl: 0.17.0
+  intl: 0.18.0
   json_path: 0.4.1
   mime_type: 1.0.0
   page_transition: 2.0.4
@@ -69,7 +70,6 @@ dependencies:
   path_provider_android: 2.0.25
   path_provider_foundation: 2.2.2
   path_provider_platform_interface: 2.0.6
-  pin_code_fields: 7.4.0
   plugin_platform_interface: 2.1.3
   provider: 6.0.4
   rxdart: 0.27.7
@@ -83,6 +83,7 @@ dependencies:
   sign_in_with_apple_web: 1.0.1
   sqflite: 2.2.6
   stream_transform: 2.1.0
+  text_search: 0.0.2
   timeago: 3.2.2
   url_launcher: 6.1.10
   url_launcher_android: 6.0.27
@@ -98,8 +99,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  syncfusion_flutter_charts: ^21.2.4
-  phone_number: ^1.0.0
 
 dev_dependencies:
   flutter_launcher_icons: 0.12.0
@@ -112,11 +111,7 @@ flutter_icons:
   android: true
   ios: true
   remove_alpha_ios: true
-  image_path: 'assets/images/logo.png'
-  web:
-    generate: true
-    image_path: 'assets/images/logo.png'
-
+  image_path: 'assets/images/app_launcher_icon.png'
 
 
 # For information on the generic Dart part of this file, see the
