@@ -3,8 +3,12 @@ import 'dart:async';
 import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+import '../auth_manager.dart';
+import '../base_auth_user_provider.dart';
+import '../../flutter_flow/flutter_flow_util.dart';
 
 import '/backend/backend.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:stream_transform/stream_transform.dart';
 import 'firebase_auth_manager.dart';
 
@@ -13,15 +17,19 @@ export 'firebase_auth_manager.dart';
 final _authManager = FirebaseAuthManager();
 FirebaseAuthManager get authManager => _authManager;
 
-String get currentUserEmail => currentUser?.email ?? '';
+String get currentUserEmail =>
+    currentUserDocument?.email ?? currentUser?.email ?? '';
 
 String get currentUserUid => currentUser?.uid ?? '';
 
-String get currentUserDisplayName => currentUser?.displayName ?? '';
+String get currentUserDisplayName =>
+    currentUserDocument?.displayName ?? currentUser?.displayName ?? '';
 
-String get currentUserPhoto => currentUser?.photoUrl ?? '';
+String get currentUserPhoto =>
+    currentUserDocument?.photoUrl ?? currentUser?.photoUrl ?? '';
 
-String get currentPhoneNumber => currentUser?.phoneNumber ?? '';
+String get currentPhoneNumber =>
+    currentUserDocument?.phoneNumber ?? currentUser?.phoneNumber ?? '';
 
 String get currentJwtToken => _currentJwtToken ?? '';
 
@@ -34,3 +42,32 @@ final jwtTokenStream = FirebaseAuth.instance
     .idTokenChanges()
     .map((user) async => _currentJwtToken = await user?.getIdToken())
     .asBroadcastStream();
+
+DocumentReference? get currentUserReference =>
+    loggedIn ? UsersRecord.collection.doc(currentUser!.uid) : null;
+
+UsersRecord? currentUserDocument;
+final authenticatedUserStream = FirebaseAuth.instance
+    .authStateChanges()
+    .map<String>((user) => user?.uid ?? '')
+    .switchMap(
+      (uid) => uid.isEmpty
+          ? Stream.value(null)
+          : UsersRecord.getDocument(UsersRecord.collection.doc(uid))
+              .handleError((_) {}),
+    )
+    .map((user) => currentUserDocument = user)
+    .asBroadcastStream();
+
+class AuthUserStreamWidget extends StatelessWidget {
+  const AuthUserStreamWidget({Key? key, required this.builder})
+      : super(key: key);
+
+  final WidgetBuilder builder;
+
+  @override
+  Widget build(BuildContext context) => StreamBuilder(
+        stream: authenticatedUserStream,
+        builder: (context, _) => builder(context),
+      );
+}
