@@ -4,14 +4,19 @@ import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import '../auth_manager.dart';
+import '../base_auth_user_provider.dart';
+import '../../flutter_flow/flutter_flow_util.dart';
 
 import '/backend/backend.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:stream_transform/stream_transform.dart';
 import 'anonymous_auth.dart';
 import 'apple_auth.dart';
 import 'email_auth.dart';
 import 'firebase_user_provider.dart';
 import 'google_auth.dart';
 import 'jwt_token_auth.dart';
+import 'github_auth.dart';
 
 export '../base_auth_user_provider.dart';
 
@@ -47,7 +52,9 @@ class FirebaseAuthManager extends AuthManager
         JwtSignInManager,
         PhoneSignInManager {
   // Set when using phone verification (after phone number is provided).
+  String? _phoneAuthVerificationCode;
   // Set when using phone sign in in web mode (ignored otherwise).
+  ConfirmationResult? _webPhoneAuthConfirmationResult;
   FirebasePhoneAuthManager phoneAuthManager = FirebasePhoneAuthManager();
 
   @override
@@ -131,6 +138,10 @@ class FirebaseAuthManager extends AuthManager
   @override
   Future<BaseAuthUser?> signInWithGoogle(BuildContext context) =>
       _signInOrCreateAccount(context, googleSignInFunc, 'GOOGLE');
+
+  @override
+  Future<BaseAuthUser?> signInWithGithub(BuildContext context) =>
+      _signInOrCreateAccount(context, githubSignInFunc, 'GITHUB');
 
   @override
   Future<BaseAuthUser?> signInWithJwtToken(
@@ -249,9 +260,12 @@ class FirebaseAuthManager extends AuthManager
   ) async {
     try {
       final userCredential = await signInFunc();
+      if (userCredential?.user != null) {
+        await maybeCreateUser(userCredential!.user!);
+      }
       return userCredential == null
           ? null
-          : AmaffFirebaseUser.fromUserCredential(userCredential);
+          : Hprms2FirebaseUser.fromUserCredential(userCredential);
     } on FirebaseAuthException catch (e) {
       ScaffoldMessenger.of(context).hideCurrentSnackBar();
       ScaffoldMessenger.of(context).showSnackBar(
