@@ -5,8 +5,8 @@ import '../base_auth_user_provider.dart';
 
 export '../base_auth_user_provider.dart';
 
-class AmaffFirebaseUser extends BaseAuthUser {
-  AmaffFirebaseUser(this.user);
+class Hprms2FirebaseUser extends BaseAuthUser {
+  Hprms2FirebaseUser(this.user);
   User? user;
   bool get loggedIn => user != null;
 
@@ -30,26 +30,31 @@ class AmaffFirebaseUser extends BaseAuthUser {
     // Reloads the user when checking in order to get the most up to date
     // email verified status.
     if (loggedIn && !user!.emailVerified) {
-      FirebaseAuth.instance.currentUser
-          ?.reload()
-          .then((_) => user = FirebaseAuth.instance.currentUser);
+      refreshUser();
     }
     return user?.emailVerified ?? false;
   }
 
+  @override
+  Future refreshUser() async {
+    await FirebaseAuth.instance.currentUser
+        ?.reload()
+        .then((_) => user = FirebaseAuth.instance.currentUser);
+  }
+
   static BaseAuthUser fromUserCredential(UserCredential userCredential) =>
       fromFirebaseUser(userCredential.user);
-  static BaseAuthUser fromFirebaseUser(User? user) => AmaffFirebaseUser(user);
+  static BaseAuthUser fromFirebaseUser(User? user) => Hprms2FirebaseUser(user);
 }
 
-Stream<BaseAuthUser> amaffFirebaseUserStream() => FirebaseAuth.instance
+Stream<BaseAuthUser> hprms2FirebaseUserStream() => FirebaseAuth.instance
         .authStateChanges()
         .debounce((user) => user == null && !loggedIn
             ? TimerStream(true, const Duration(seconds: 1))
             : Stream.value(user))
         .map<BaseAuthUser>(
       (user) {
-        currentUser = AmaffFirebaseUser(user);
+        currentUser = Hprms2FirebaseUser(user);
         return currentUser!;
       },
     );
