@@ -1,4 +1,6 @@
-export 'unique_crop_name.dart' show uniqueCropName;
-export 'unique_phone_number.dart' show uniquePhoneNumber;
-export 'append_country_code.dart' show appendCountryCode;
-export 'get_weather_settings.dart' show getWeatherSettings;
+export 'create_doctor.dart' show createDoctor;
+export 'get_filtered_doctor_list.dart' show getFilteredDoctorList;
+export 'save_my_schedule.dart' show saveMySchedule;
+export 'get_doctor_ref_by_name.dart' show getDoctorRefByName;
+export 'check_if_lab_result_exists.dart' show checkIfLabResultExists;
+export 'send_email_notif.dart' show sendEmailNotif;
