@@ -1,11 +1,25 @@
 // Export pages
-export '/forgot_password/forgot_password_widget.dart' show ForgotPasswordWidget;
-export '/phone_sign_in/phone_sign_in_widget.dart' show PhoneSignInWidget;
-export '/phone_verify/phone_verify_widget.dart' show PhoneVerifyWidget;
-export '/users/users_widget.dart' show UsersWidget;
-export '/crops_page/crops_page_widget.dart' show CropsPageWidget;
-export '/home_page/home_page_widget.dart' show HomePageWidget;
-export '/customers_page/customers_page_widget.dart' show CustomersPageWidget;
-export '/create_user/create_user_widget.dart' show CreateUserWidget;
-export '/create_crop/create_crop_widget.dart' show CreateCropWidget;
-export '/settings_page/settings_page_widget.dart' show SettingsPageWidget;
+export '/pages/login_page/login_page_widget.dart' show LoginPageWidget;
+export '/pages/complete_profile/complete_profile_widget.dart'
+    show CompleteProfileWidget;
+export '/pages/forgot_password/forgot_password_widget.dart'
+    show ForgotPasswordWidget;
+export '/pages/add_another_profile/add_another_profile_widget.dart'
+    show AddAnotherProfileWidget;
+export '/pages/home_page/home_page_widget.dart' show HomePageWidget;
+export '/pages/my_appointments/my_appointments_widget.dart'
+    show MyAppointmentsWidget;
+export '/pages/appointment_details/appointment_details_widget.dart'
+    show AppointmentDetailsWidget;
+export '/pages/profile_page/profile_page_widget.dart' show ProfilePageWidget;
+export '/pages/edit_profile/edit_profile_widget.dart' show EditProfileWidget;
+export '/pages/find_symptoms/find_symptoms_widget.dart' show FindSymptomsWidget;
+export '/pages/appointment_details_profile/appointment_details_profile_widget.dart'
+    show AppointmentDetailsProfileWidget;
+export '/pages/doctors_page/doctors_page_widget.dart' show DoctorsPageWidget;
+export '/pages/edit_doctor/edit_doctor_widget.dart' show EditDoctorWidget;
+export '/pages/patients_page/patients_page_widget.dart' show PatientsPageWidget;
+export '/pages/my_schedule/my_schedule_widget.dart' show MyScheduleWidget;
+export '/pages/fecalysis_form/fecalysis_form_widget.dart'
+    show FecalysisFormWidget;
+export '/pages/dental_form/dental_form_widget.dart' show DentalFormWidget;
